Update for MassDigi Challenge
- Added background effects and particles to main menu and to all the
levels
- Changed colors of the main menu buttons and panels
- Added music to the main menu
- Added bits of music to the levels and each time you die a track will
be chosen randomly to be played. This avoids the repetition of
restarting the music from the beginning every single time.
- Added rewind sound effects, when you die the music will stop, when you
click on the screen the rewind effect will be played and will loop until
you go back to the beginning of the level
- Changed the enemies with the new model, but kept the old model for the
patrolling enemies
- Fixed the lowest horizontal collision problem, now the player no
longer goes through it